Refactor: Split ChiExtractionTagger into modular components
